Update bahis_dash.py
Several bugfixes, but not the double names yet.
Disease Case can now show the same as the first "expander" with viewing dynamics with the slight difference in the line chart appearance.
In principle, I would deactivate the first expander and the third with "Upazila data" and move the overview behind the disease cases.
Rename disease cases.